Additional multifidelity bug fixes from testing on multilevel_harden
1. Reset the integration driver for multifidelity UQ including refinement.
2. Correct overly-aggressive pruning of full covariance stats.
3. Correct use of unsigned short instead of size_t for collocation pts in
   config_regression().
4. Add missing "#' in dakota_sbuq_hier_poly_prod.in
5. Bless more compact results reports for dakota_sbuq_*.in, which are now clean.

+++ src/NonDPolynomialChaos.cpp
@@ -481,7 +481,7 @@ config_expectation(size_t exp_samples, unsigned short sample_type,
 
 
 bool NonDPolynomialChaos::
-config_regression(const UShortArray& exp_orders, unsigned short colloc_pts,
+config_regression(const UShortArray& exp_orders, size_t colloc_pts,
 		  Real colloc_ratio_terms_order, short regress_type,
 		  short ls_regress_type, const UShortArray& tensor_grid_order,
 		  unsigned short sample_type, const String& rng,
